backend/requirements.txt
devtools
openai
opencv-python
python-dotenv